delete old bot file and rename new one to main.py
